+++ packages/webapp/src/pages/Integrations/components/forms/AuthCreateForm.tsx
@@ -1,11 +1,10 @@
 import { AppAuthCreateForm } from './AppAuthCreateForm';
 import { CustomAuthCreateForm } from './CustomAuthCreateForm';
+import { DefaultCreateForm } from './DefaultCreateForm';
 import { InstallPluginAuthCreateForm } from './InstallPluginAuthCreateForm';
 import { McpGenericCreateForm } from './McpGenericCreateForm';
 import { McpOAuthCreateForm } from './McpOAuthCreateForm';
 import { OAuthCreateForm } from './OAuthCreateForm';
-import { Alert, AlertDescription } from '@/components-v2/ui/alert';
-import { Button } from '@/components-v2/ui/button';
 
 import type { ApiProviderListItem, PostIntegration } from '@nangohq/types';
 
@@ -13,17 +12,6 @@ interface Props {
     provider: ApiProviderListItem;
     onSubmit?: (data: PostIntegration['Body']) => Promise<void>;
 }
-
-const getInfoMessage = (provider: ApiProviderListItem): string | null => {
-    switch (provider.authMode) {
-        case 'BASIC':
-            return "This API uses basic auth. Nothing to configure here, Nango will ask for the user's basic credentials as part of the auth flow.";
-        case 'API_KEY':
-            return 'This API uses API key auth. Nothing to configure here, Nango will ask the user for an API key as part of the auth flow.';
-    }
-
-    return `Nothing to configure here.`;
-};
 
 export const AuthCreateForm: React.FC<Props> = ({ provider, onSubmit }) => {
     const authMode = provider.authMode;
@@ -49,28 +37,7 @@ export const AuthCreateForm: React.FC<Props> = ({ provider, onSubmit }) => {
         case 'INSTALL_PLUGIN':
             return <InstallPluginAuthCreateForm provider={provider} onSubmit={onSubmit} />;
 
-        default: {
-            const infoMessage = getInfoMessage(provider);
-            return (
-                <div className="flex flex-col gap-8">
-                    {infoMessage && (
-                        <Alert variant="info">
-                            <AlertDescription>{infoMessage}</AlertDescription>
-                        </Alert>
-                    )}
-                    <Button
-                        variant="primary"
-                        onClick={() =>
-                            onSubmit?.({
-                                provider: provider.name,
-                                useSharedCredentials: false
-                            })
-                        }
-                    >
-                        Create
-                    </Button>
-                </div>
-            );
-        }
+        default:
+            return <DefaultCreateForm provider={provider} onSubmit={onSubmit} />;
     }
 };
